refactor tests to use the microtest framework
