+++ test/sql_asserts.h
